better search mode logic
